requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
frozenlist==1.5.0
gotrue==2.11.4
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.9.0
multidict==6.2.0
openai==1.66.3
packaging==24.2
passlib==1.7.4
postgrest==0.19.3
propcache==0.3.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
realtime==2.4.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlmodel==0.0.24
starlette==0.46.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.14.0
supafunc==0.9.3
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.2
yarl==1.18.3
